requirements.txt
# Deep Learning Framework
tensorflow==2.9.0  # CUDA 11.2 compatible

# Scientific Computing
numpy>=1.21.0,<1.24.0  # TF 2.9 compatibility
pandas>=1.3.0
scipy>=1.7.0

# Machine Learning
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0

# Data Processing
pillow>=10.0.0
opencv-python>=4.8.0

# Configuration & Logging
pyyaml>=6.0
tqdm>=4.65.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Code Quality
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0

# API Service
flask>=2.3.0

# Utilities
python-dotenv>=1.0.0
pathlib>=1.0.1
